chore(dotnet): s/Readme/ReadMe across codebase (#525)
* chore(dotnet): s/Readme/ReadMe across codebase

I believe packages on NuGet are case insensitive, and luckily the casing
is already correct on there: https://www.nuget.org/packages/ReadMe.Metrics/

Lets see if the tests pass here 🤞

* chore(dotnet): remove solution file from source control

Since I dont use Visual Studio, I dont think we need this?

* chore(dotnet): rename Readme folder to ReadMe

* fix(dotnet): rename Readme.csproj to ReadMe.csproj

* chore(prettier): fix markdown and update prettierignore

* chore(dotnet): update last remaining references to Readme
