+++ packages/dotnet/ReadMe/HarJsonObjectModels/Cookies.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace ReadMe.HarJsonObjectModels
+{
+  [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
+  class Cookies
+  {
+    public string name { get; set; }
+    public string value { get; set; }
+  }
+}
