+++ packages/dotnet/ReadMe/HarJsonObjectModels/Group.cs
@@ -1,0 +1,10 @@
+﻿namespace ReadMe.HarJsonObjectModels
+{
+  public class Group
+  {
+    public string email { get; set; }
+    public string label { get; set; }
+    public string id { get; set; }
+
+  }
+}
