+++ packages/dotnet/ReadMe/HarJsonObjectModels/Headers.cs
@@ -1,0 +1,8 @@
+﻿namespace ReadMe.HarJsonObjectModels
+{
+  class Headers
+  {
+    public string name { get; set; }
+    public string value { get; set; }
+  }
+}
