+++ packages/dotnet/ReadMe/HarJsonObjectModels/Log.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace ReadMe.HarJsonObjectModels
+{
+  class Log
+  {
+    public Creator creator { get; set; }
+    public List<Entries> entries { get; set; }
+
+  }
+}
