+++ packages/dotnet/ReadMe/HarJsonObjectModels/RequestMain.cs
@@ -1,0 +1,12 @@
+﻿namespace ReadMe.HarJsonObjectModels
+{
+  class RequestMain
+  {
+    public Log log { get; set; }
+
+    public RequestMain(Log log)
+    {
+      this.log = log;
+    }
+  }
+}
