+++ packages/dotnet/ReadMe/HarJsonObjectModels/Timing.cs
@@ -1,0 +1,8 @@
+﻿namespace ReadMe.HarJsonObjectModels
+{
+  class Timing
+  {
+    public int wait { get; set; }
+    public string receive { get; set; }
+  }
+}
