--- packages/dotnet/Readme/HarJsonObjectModels/Creator.cs
@@ -1,15 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-
-namespace Readme.HarJsonObjectModels
-{
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
-    class Creator
-    {
-        public string name { get; set; }
-        public string version { get; set; }
-        //comment is OS and its version
-        public string comment { get; set; }
-
-    }
-}
