--- packages/dotnet/Readme/HarJsonObjectModels/Group.cs
@@ -1,10 +1,0 @@
-﻿namespace Readme.HarJsonObjectModels
-{
-    public class Group
-    {
-        public string email { get; set; }
-        public string label { get; set; }
-        public string id { get; set; }
-
-    }
-}
