--- packages/dotnet/Readme/HarJsonObjectModels/Headers.cs
@@ -1,8 +1,0 @@
-﻿namespace Readme.HarJsonObjectModels
-{
-    class Headers
-    {
-        public string name { get; set; }
-        public string value { get; set; }
-    }
-}
