--- packages/dotnet/Readme/HarJsonObjectModels/Log.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Readme.HarJsonObjectModels
-{
-    class Log
-    {
-        public Creator creator { get; set; }
-        public List<Entries> entries { get; set; }
-
-    }
-}
