--- packages/dotnet/Readme/HarJsonObjectModels/Params.cs
@@ -1,14 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace Readme.HarJsonObjectModels
-{
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
-    class Params
-    {
-        public string name { get; set; }
-        public string value { get; set; }
-        public string fileName { get; set; }
-        public string contentType { get; set; }
-        public string comment { get; set; }
-    }
-}
