--- packages/dotnet/Readme/HarJsonObjectModels/Request.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Readme.HarJsonObjectModels
-{
-    class Request
-    {
-        public List<Headers> headers { get; set; }
-        public long headersSize { get; set; }
-        public List<QueryStrings> queryString { get; set; }
-
-        public PostData postData { get; set; }
-        public List<Cookies> cookies { get; set; }
-        public string method { get; set; }
-        public string url { get; set; }
-        public string httpVersion { get; set; }
-    }
-}
