--- packages/dotnet/Readme/HarJsonObjectModels/RequestMain.cs
@@ -1,12 +1,0 @@
-﻿namespace Readme.HarJsonObjectModels
-{
-    class RequestMain
-    {
-        public Log log { get; set; }
-
-        public RequestMain(Log log)
-        {
-            this.log = log;
-        }
-    }
-}
