--- packages/dotnet/Readme/HarJsonObjectModels/Response.cs
@@ -1,19 +1,0 @@
-﻿using Newtonsoft.Json;
-using System.Collections.Generic;
-
-namespace Readme.HarJsonObjectModels
-{
-    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
-    class Response
-    {
-        public List<Headers> headers { get; set; }
-        public Content content { get; set; }
-        public int status { get; set; }
-        public string statusText { get; set; }
-        public string httpVersion { get; set; }
-        public List<Cookies> cookies { get; set; }
-        public string redirectURL { get; set; }
-        public long headersSize { get; set; }
-        public long bodySize { get; set; }
-    }
-}
