--- packages/dotnet/Readme/HarJsonObjectModels/Root.cs
@@ -1,12 +1,0 @@
-﻿namespace Readme.HarJsonObjectModels
-{
-    class Root
-    {
-        public string _id { get; set; }
-        public bool development { get; set; }
-        public string clientIPAddress { get; set; }
-        public Group group { get; set; }
-        public RequestMain request { get; set; }
-
-    }
-}
