--- packages/dotnet/Readme/HarJsonObjectModels/Timing.cs
@@ -1,8 +1,0 @@
-﻿namespace Readme.HarJsonObjectModels
-{
-    class Timing
-    {
-        public int wait { get; set; }
-        public string receive { get; set; }
-    }
-}
