--- packages/dotnet/ReadmeMetrics.sln
@@ -1,49 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32014.148
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReadmeMetricsAPI6", "ReadmeMetricsAPI6\ReadmeMetricsAPI6.csproj", "{7BEC2951-B728-41FC-BE08-A75F1F49F483}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D55C05A3-271F-4E88-A029-764DE9AED8F2}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-		readme2.png = readme2.png
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReadmeMetricsAPI5", "ReadmeMetricsAPI5\ReadmeMetricsAPI5.csproj", "{51CF7339-D4CF-43AB-A3E5-D42ADAD309B1}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReadmeMetricsAPI3", "ReadmeMetricsAPI3\ReadmeMetricsAPI3.csproj", "{DA88FA73-C226-4958-ACDD-E8BD8E0BAECD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Readme", "Readme\Readme.csproj", "{48A3FD74-06DA-48FC-85EC-C2CD90008537}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7BEC2951-B728-41FC-BE08-A75F1F49F483}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BEC2951-B728-41FC-BE08-A75F1F49F483}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BEC2951-B728-41FC-BE08-A75F1F49F483}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BEC2951-B728-41FC-BE08-A75F1F49F483}.Release|Any CPU.Build.0 = Release|Any CPU
-		{51CF7339-D4CF-43AB-A3E5-D42ADAD309B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{51CF7339-D4CF-43AB-A3E5-D42ADAD309B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{51CF7339-D4CF-43AB-A3E5-D42ADAD309B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{51CF7339-D4CF-43AB-A3E5-D42ADAD309B1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DA88FA73-C226-4958-ACDD-E8BD8E0BAECD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA88FA73-C226-4958-ACDD-E8BD8E0BAECD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA88FA73-C226-4958-ACDD-E8BD8E0BAECD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA88FA73-C226-4958-ACDD-E8BD8E0BAECD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{48A3FD74-06DA-48FC-85EC-C2CD90008537}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{48A3FD74-06DA-48FC-85EC-C2CD90008537}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{48A3FD74-06DA-48FC-85EC-C2CD90008537}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{48A3FD74-06DA-48FC-85EC-C2CD90008537}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {497D459C-5910-45C7-811D-312B2584F971}
-	EndGlobalSection
-EndGlobal
