+++ packages/dotnet/examples/net6.0-webhook/Program.cs
@@ -21,7 +21,7 @@ app.MapPost("/webhook", async context =>
 
   try
   {
-    Readme.Webhook.Verify(body, signature, secret);
+    ReadMe.Webhook.Verify(body, signature, secret);
   }
   catch (Exception e)
   {
