+++ packages/dotnet/examples/net6.0/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "readme": {
-    "apiKey": "<Your Readme API Key>"
+    "apiKey": "<Your ReadMe API Key>"
   }
 }
